feat(FX-4772): Inline price control with Create alert modal (#12467)
* feat(FX-4772): Inline price control with Create alert modal

* Refactoring WIP

* Refactoring WIP

* Bugfixes

* Code review fixes

* Fixes type mismatch

* Temp

* Fixes folder capitalization

* translation
