+++ src/Components/PriceRange/Histogram.tsx
