+++ src/Components/PriceRange/Utils/convertToFilterFormatRange.ts
@@ -1,0 +1,11 @@
+import { DEFAULT_RANGE } from "Components/PriceRange/constants"
+
+export const convertToFilterFormatRange = (range: number[]) => {
+  return range.map((value, index) => {
+    if (value === DEFAULT_RANGE[index]) {
+      return "*"
+    }
+
+    return value
+  })
+}
