+++ src/Components/PriceRange/Utils/getPriceValue.ts
@@ -1,0 +1,5 @@
+import { CustomRange } from "Components/PriceRange/constants"
+
+export const getPriceValue = (value: CustomRange[number]) => {
+  return value === "*" || value === 0 ? "" : value
+}
