+++ src/Components/SavedSearchAlert/SavedSearchAlertModal.tsx
@@ -1,4 +1,4 @@
-import React from "react"
+import React, { FC } from "react"
 import { Formik } from "formik"
 import {
   Box,
@@ -8,6 +8,7 @@ import {
   Input,
   Join,
   ModalDialog,
+  Separator,
   Spacer,
   Text,
 } from "@artsy/palette"
@@ -31,6 +32,7 @@ import { SavedSearchAlertPills } from "./Components/SavedSearchAlertPills"
 import { Metric } from "Utils/metrics"
 import { DEFAULT_FREQUENCY } from "./constants"
 import { FrequenceRadioButtons } from "./Components/FrequencyRadioButtons"
+import { PriceRangeFilter } from "Components/SavedSearchAlert/Components/PriceRangeFilter"
 
 interface SavedSearchAlertFormProps {
   entity: SavedSearchEntity
@@ -49,7 +51,7 @@ export interface SavedSearchAlertFormContainerProps
 
 const logger = createLogger("Components/SavedSearchAlert/SavedSearchAlertModal")
 
-export const SavedSearchAlertModal: React.FC<SavedSearchAlertFormProps> = ({
+export const SavedSearchAlertModal: FC<SavedSearchAlertFormProps> = ({
   entity,
   initialValues,
   onClose,
@@ -84,7 +86,6 @@ export const SavedSearchAlertModal: React.FC<SavedSearchAlertFormProps> = ({
         userAlertSettings,
         criteria
       )
-
       const result = {
         id: response.createSavedSearch?.savedSearchOrErrors.internalID!,
       }
@@ -126,7 +127,7 @@ export const SavedSearchAlertModal: React.FC<SavedSearchAlertFormProps> = ({
               </Button>
             }
           >
-            <Join separator={<Spacer y={4} />}>
+            <Join separator={<Spacer y={2} />}>
               <Input
                 title="Alert Name"
                 name="name"
@@ -147,6 +148,12 @@ export const SavedSearchAlertModal: React.FC<SavedSearchAlertFormProps> = ({
                     onDeletePress={handleRemovePillPress}
                   />
                 </Flex>
+
+                <Separator my={1} />
+
+                <PriceRangeFilter />
+
+                <Separator my={2} />
               </Box>
 
               <Box>
@@ -173,7 +180,7 @@ export const SavedSearchAlertModal: React.FC<SavedSearchAlertFormProps> = ({
                   />
                 </Box>
 
-                <Spacer y={4} />
+                <Spacer y={2} />
 
                 {values.push && (
                   <FrequenceRadioButtons
